refactor(code-drawing): migrate to register
