+++ apps/www/src/registry/components/editor/plugins/code-drawing-base-kit.tsx
@@ -1,11 +1,9 @@
-'use client';
-
-import { toPlatePlugin } from 'platejs/react';
-
 import { BaseCodeDrawingPlugin } from '@platejs/code-drawing';
 
 import { CodeDrawingElement } from '@/registry/ui/code-drawing-node';
 
 export const BaseCodeDrawingKit = [
-  toPlatePlugin(BaseCodeDrawingPlugin).withComponent(CodeDrawingElement),
+  BaseCodeDrawingPlugin.configure({
+    node: { component: CodeDrawingElement },
+  }),
 ];
