--- packages/code-drawing/src/react/components/CodeDrawingFloatingToolbar.tsx
@@ -1,58 +1,0 @@
-'use client';
-
-import * as React from 'react';
-
-import type { TCodeDrawingElement } from '../../lib';
-
-import {
-  useEditorSelector,
-  useFocusedLast,
-  useReadOnly,
-  useSelected,
-} from 'platejs/react';
-
-export type CodeDrawingFloatingToolbarProps = {
-  children: React.ReactNode;
-  element: TCodeDrawingElement;
-  onRemove: () => void;
-  onDownload?: () => void;
-  renderPopover?: (props: {
-    children: React.ReactNode;
-    onRemove: () => void;
-    onDownload?: () => void;
-    open: boolean;
-  }) => React.ReactNode;
-};
-
-export function CodeDrawingFloatingToolbar({
-  children,
-  element: _element,
-  onRemove,
-  onDownload,
-  renderPopover,
-}: CodeDrawingFloatingToolbarProps) {
-  const readOnly = useReadOnly();
-  const selected = useSelected();
-  const isFocusedLast = useFocusedLast();
-  const selectionCollapsed = useEditorSelector(
-    (editor) => !editor.api.isExpanded(),
-    []
-  );
-
-  const open = isFocusedLast && !readOnly && selected && selectionCollapsed;
-
-  if (!renderPopover) {
-    return <>{children}</>;
-  }
-
-  return (
-    <>
-      {renderPopover({
-        children,
-        onRemove,
-        onDownload,
-        open,
-      })}
-    </>
-  );
-}
