--- packages/code-drawing/src/react/components/index.ts
@@ -1,9 +1,0 @@
-/**
- * @file Automatically generated by barrelsby.
- */
-
-export * from './CodeDrawingFloatingToolbar';
-export * from './CodeDrawingPreview';
-export * from './CodeDrawingPreviewArea';
-export * from './CodeDrawingTextarea';
-export * from './CodeDrawingToolbar';
